+++ prr-core/src/prr/communications/InteractiveCommunication.java
@@ -8,7 +8,7 @@ public abstract class InteractiveCommunication extends Communication {
         super(key, sender, receiver, 0);
     }
 
-    public long endCommunication(int duration, boolean isFriend) {
+    public double endCommunication(int duration, boolean isFriend) {
         this.finish();
         this.setUnits(duration);
         this.updatePrice(isFriend);
